messaging-service/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
fastapi==0.104.1
uvicorn==0.24.0
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4
structlog==23.1.0
tenacity==8.2.3
jsonschema==4.19.1
flask==3.0.0
flask-cors==4.0.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
kafka-python-ng==2.2.0